+++ config/src/lib.rs
@@ -1,6 +1,5 @@
 //! Provides a general solution to maintain configuration spanned across different sources.
 //!
-//! The configuration options can be set in various levels of increasing priority:
 //!
 //! - Default Values
 //! - System configuration
@@ -37,7 +36,7 @@ use std::{
     path::{Path, PathBuf},
     str::FromStr,
 };
-use xvc_logging::watch;
+use xvc_logging::{debug, watch};
 use xvc_walker::AbsolutePath;
 
 use strum_macros::{Display as EnumDisplay, EnumString, IntoStaticStr};