+++ core/src/types/xvcpath.rs
@@ -248,7 +248,7 @@ impl XvcCachePath {
 
     /// Returns the prefix for this digest for reporting purposes
     /// len is the number of digits to return from
-    pub fn digest_prefix(&self, len: usize) -> String {
+    pub fn digest_string(&self, len: usize) -> String {
         self.0
             .to_string()
             .chars()