+++ file/src/untrack/mod.rs
@@ -1,5 +1,6 @@
-use std::collections::HashSet;
-use std::fs::create_dir_all;
+use std::collections::{HashMap, HashSet};
+use std::fs::{self, create_dir_all};
+use std::path::PathBuf;
 
 use crate::common::gitignore::make_ignore_handler;
 use crate::common::{
@@ -11,8 +12,12 @@ use clap::Parser;
 use crossbeam_channel::Sender;
 use derive_more::From;
 use itertools::Itertools;
-use xvc_core::types::cachetype;
-use xvc_core::{CacheType, ContentDigest, XvcCachePath, XvcMetadata, XvcRoot};
+use rayon::prelude::{IntoParallelRefIterator, ParallelIterator};
+use xvc_core::types::recheckmethod;
+use xvc_core::{ContentDigest, RecheckMethod, XvcCachePath, XvcMetadata, XvcPath, XvcRoot};
+use xvc_ecs::ecs::event::EventLog;
+use xvc_ecs::{HStore, XvcEntity, XvcStore};
+use xvc_logging::{debug, error, output, panic, uwo, uwr, watch};
 
 /// Remove files from tracking and possibly delete them
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, From, Parser)]
@@ -45,19 +50,21 @@ pub fn cmd_untrack(
 
     // Get cache paths for each
 
-    let all_cache_paths: HStore<Vec<XvcCachePath>> = HStore::new();
+    let mut all_cache_paths: HStore<Vec<XvcCachePath>> = HStore::new();
 
     // Find all cache paths
     // We have 1-1 relationship between content digests and paths.
     // So, in order to get earlier versions, we check the event log.
     for (xe, xp) in all_paths.iter() {
         let path_digest_events: EventLog<ContentDigest> =
-            all_content_digests.all_event_log_for_entity(xe)?;
+            all_content_digests.all_event_log_for_entity(*xe)?;
         let cache_paths = path_digest_events
             .iter()
             .filter_map(|cd_event| match cd_event {
                 xvc_ecs::ecs::event::Event::Add { entity, value } => {
-                    let xcp = XvcCachePath::new(xp, value)?;
+                    let xcp = uwr!(XvcCachePath::new(xp, value), output_snd
+                 );
+
                     Some(xcp)
                 }
                 xvc_ecs::ecs::event::Event::Remove { entity } => {
@@ -77,15 +84,15 @@ pub fn cmd_untrack(
 
     // Get inverted index
 
-    let mut entities_for_cache_paths: HashMap<XvcCachePath, HashSet<XvcEntity>> = HashMap::new();
+    let mut entities_for_cache_path: HashMap<XvcCachePath, HashSet<XvcEntity>> = HashMap::new();
 
     for (xe, cache_paths) in all_cache_paths.iter() {
         for cp in cache_paths {
-            if !entities_for_cache_paths.contains_key(cp) {
-                entities_for_cache_paths.insert(cp, HashSet::new());
+            if !entities_for_cache_path.contains_key(cp) {
+                entities_for_cache_path.insert(cp.clone(), HashSet::new());
             }
-            let mut entity_set = entities_for_cache_paths.get_mut(cp).unwrap();
-            entity_set.push(*xe);
+            let mut entity_set = entities_for_cache_path.get_mut(cp).unwrap();
+            entity_set.insert(*xe);
         }
     }
 
@@ -93,39 +100,39 @@ pub fn cmd_untrack(
 
     let (ignore_snd, ignore_thread) = make_ignore_handler(output_snd, xvc_root)?;
     let (recheck_op_snd, recheck_thread) = make_recheck_handler(output_snd, xvc_root, &ignore_snd)?;
-    let all_recheck_methods = xvc_root.load_store::<CacheType>()?;
-    let target_recheck_methods = all_recheck_methods.subset(untrack_targets.keys())?;
 
-    target_recheck_methods.iter().for_each(|(xe, rm)| match rm {
-        CacheType::Copy | CacheType::Hardlink => {
-            let xp = all_paths[xe];
-            debug!(output_snd, "Path is already copied: {xp}");
+    for (xe, xvc_path) in untrack_targets.iter() {
+        let abs_path = xvc_path.to_absolute_path(xvc_root);
+        if !abs_path.is_file() {
+            let content_digest = all_content_digests[&xe];
+            let recheck_method = RecheckMethod::Copy;
+            uwr!(
+                recheck_op_snd.send(Some(RecheckOperation::Recheck {
+                    xvc_path: xvc_path.clone(),
+                    content_digest,
+                    recheck_method,
+                })),
+                output_snd
+            );
+        } else {
+            debug!(output_snd, "Path is already a file: {xvc_path}");
         }
-        // If the underlying fs doesn't support reflinks, this isn't required as we fall back to copy.
-        // But this looks premature optimization now.
-        CacheType::Reflink | CacheType::Symlink => {
-            let xvc_path = all_paths[xe];
-            let content_digest = all_content_digests[xe];
-            let recheck_method = CacheType::Copy;
-            recheck_op_snd.send(Some(RecheckOperation {
-                xvc_path,
-                content_digest,
-                cache_type: recheck_method,
-            }))?;
-        }
-    });
+    }
 
-    recheck_op_snd.send(None)?;
-    recheck_thread.join().unwrap();
+    uwr!(recheck_op_snd.send(None), output_snd);
+    uwr!(recheck_thread.join(), output_snd);
+    uwr!(ignore_snd.send(None), output_snd);
+    uwr!(ignore_thread.join(), output_snd);
 
     // Find cache paths of the targets
 
     let untrack_cache_paths: HStore<Vec<XvcCachePath>> =
-        all_cache_paths.subset(untrack_targets.keys())?;
-    let untrack_entities = HashSet::from(untrack_targets.keys());
+        all_cache_paths.subset(untrack_targets.keys().copied())?;
+    let untrack_entities: HashSet<XvcEntity> = HashSet::from_iter(untrack_targets.keys().copied());
 
     // We can restore versions here
     if let Some(restore_dir) = opts.restore_versions {
+        const VERSION_ID_LEN: usize = 15;
         let abs_restore_dir = current_dir.join(restore_dir);
         watch!(abs_restore_dir);
         if abs_restore_dir.is_file() {
@@ -140,7 +147,7 @@ pub fn cmd_untrack(
         }
 
         untrack_cache_paths.par_iter().for_each(|(xe, vec_cp)| {
-            let xvc_path = all_paths[xe];
+            let xvc_path = all_paths[xe].clone();
             let destination_path = xvc_path.to_absolute_path(&abs_restore_dir);
             let destination_dir = destination_path.parent().unwrap();
             let stem = xvc_path
@@ -153,33 +160,35 @@ pub fn cmd_untrack(
                 .unwrap_or_else(|| "");
 
             if !destination_dir.exists() {
-                fs::create_dir_all(&destination_dir)?;
+                uwr!(fs::create_dir_all(&destination_dir), output_snd);
             }
 
             vec_cp.par_iter().for_each(|xcp| {
                 let from = xcp.to_absolute_path(xvc_root);
-                // FIXME: This is hardcoded, sorry
-                // It will be like b3-123-456-789
-                let xcp_suffix = xcp.digest_prefix(15);
+                // The suffix will be like b3-123-456-789
+                let xcp_suffix = xcp.digest_string(VERSION_ID_LEN);
                 // The file name will be like {restore_dir}/{xvc_path_parent}/{xvc_path_stem}-{cache_prefix}.{extension}
-                let to =
-                    PathBuf::from(format!("{destination_dir}/{stem}-{xcp_suffix}.{extension}"));
-                fs::copy(&from, to)?;
-                output!(output_snd, "[COPY] {from} -> {to}");
+                let to = PathBuf::from(format!(
+                    "{destination_dir}/{stem}-{xcp_suffix}.{extension}",
+                    destination_dir = destination_dir.to_str().unwrap()
+                ));
+                uwr!(fs::copy(&from, &to), output_snd);
+                output!(output_snd, "[COPY] {from} -> {}", to.to_str().unwrap());
             });
         })
     }
 
-    let mut deletable_paths: Vec<XvcCachePath>::new();
+    let mut deletable_paths = Vec::<XvcCachePath>::new();
     // Report the differences if found
     for (xe, vec_cp) in untrack_cache_paths {
         for cp in vec_cp {
-            let entities_pointing_to_cp = HashSet::from(entities_for_cache_path[cp]);
+            let entities_pointing_to_cp =
+                HashSet::from_iter(entities_for_cache_path[&cp].iter().copied());
             let mut deletable = true;
             entities_pointing_to_cp
                 .difference(&untrack_entities)
                 .for_each(|other_xe| {
-                    let this_xp = all_paths.get(xe).unwrap();
+                    let this_xp = all_paths.get(&xe).unwrap();
                     let other_xp = all_paths.get(other_xe).unwrap();
                     output!(
                         output_snd,
@@ -201,35 +210,35 @@ pub fn cmd_untrack(
     // We use nested stores to make this transactional
     xvc_root.with_store_mut(|xp_store: &mut XvcStore<XvcPath>| {
         xvc_root.with_store_mut(|xmd_store: &mut XvcStore<XvcMetadata>| {
-            xvc_root.with_store_mut(|cache_type_store: &mut XvcStore<CacheType>| {
+            xvc_root.with_store_mut(|cache_type_store: &mut XvcStore<RecheckMethod>| {
                 xvc_root.with_store_mut(
                     |text_or_binary_store: &mut XvcStore<FileTextOrBinary>| {
                         xvc_root.with_store_mut(
                             |content_digest_store: &mut XvcStore<ContentDigest>| {
-                                for xe in untrack_entities {
-                                    content_digest_store.remove(xe);
+                                for xe in &untrack_entities {
+                                    content_digest_store.remove(*xe);
                                 }
                                 Ok(())
                             },
                         )?;
-                        for xe in untrack_entities {
-                            text_or_binary_store.remove(xe);
+                        for xe in &untrack_entities {
+                            text_or_binary_store.remove(*xe);
                         }
                         Ok(())
                     },
                 )?;
-                for xe in untrack_entities {
-                    cache_type_store.remove(xe);
+                for xe in &untrack_entities {
+                    cache_type_store.remove(*xe);
                 }
                 Ok(())
             })?;
-            for xe in untrack_entities {
-                xmd_store.remove(xe);
+            for xe in &untrack_entities {
+                xmd_store.remove(*xe);
             }
             Ok(())
         })?;
-        for xe in untrack_entities {
-            xmd_store.remove(xe);
+        for xe in &untrack_entities {
+            xp_store.remove(*xe);
         }
         Ok(())
     })?;